+++ lib/src/types/address.dart
@@ -1,8 +1,8 @@
 import 'package:solana_dart/src/base58/base58.dart' as base58;
 import 'package:solana_dart/src/types/serializable.dart';
 
-class SerializableAddress extends Serializable {
-  SerializableAddress.from(this._address);
+class Address extends Serializable {
+  Address.from(this._address);
 
   final String _address;