+++ lib/src/types/tx_signature.dart
@@ -1,11 +1,12 @@
-import 'dart:convert';
+import 'package:solana_dart/src/types/common_tx.dart';
+import 'package:solana_dart/src/types/json_rpc_response_object.dart';
 
-class TxSignature {
-  TxSignature._(dynamic value) : _signature = value['result'];
+class TxSignature extends CommonTx {
+  TxSignature._(dynamic value) : _signature = value;
   TxSignature.fromJsonRpcResponseString(jsonRpcResponseString)
-      : this._(
-          json.decode(jsonRpcResponseString),
-        );
-  String toString() => _signature;
+      : this._(JsonRpcResponseObject.getResult(jsonRpcResponseString));
+
   final String _signature;
+
+  String toString() => _signature;
 }